move scripts in folder
